Extracts ApmServerRule for use in Servlet integration tests
Before we migrate to jupiter (e.g. via openrewrite), we should have code organized as typical JUnit 4 as possible. This extracts out static initialization in servlet ITs as a class rule while keeping all other logic the same. Next steps will attempt to make `TestApp` more normalized, possibly via rules or nested or something similar, I'm not quite sure yet, but the goal is to be able to have tests be normal junit tests vs loops.
