--- integration-tests/application-server-integration-tests/src/test/java/co/elastic/apm/servlet/MockServerContainer.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package co.elastic.apm.servlet;
-
-import com.github.dockerjava.api.command.InspectContainerResponse;
-import org.mockserver.client.MockServerClient;
-import org.testcontainers.containers.GenericContainer;
-
-/**
- * Acts as a mock for the APM-server.
- */
-public class MockServerContainer extends GenericContainer<MockServerContainer> {
-
-    public static final String HEALTH_ENDPOINT = "/mockserver/status";
-    private MockServerClient client;
-
-    public MockServerContainer() {
-        super("mockserver/mockserver:5.14.0");
-        addEnv("MOCKSERVER_LIVENESS_HTTP_GET_PATH", HEALTH_ENDPOINT);
-        addExposedPorts(1080);
-    }
-
-    @Override
-    protected void containerIsStarted(InspectContainerResponse containerInfo) {
-        super.containerIsStarted(containerInfo);
-        client = new MockServerClient(getContainerIpAddress(), getFirstMappedPort());
-    }
-
-    public MockServerClient getClient() {
-        return client;
-    }
-}
